requirements.txt
amqp==2.2.2
anyjson==0.3.3
billiard==3.5.0.3
cairocffi==0.8.0
CairoSVG==1.0.22
celery==4.1.0
cffi==1.10.0
cssselect==1.0.1
Django==1.10
django-appconf==1.0.2
django-auditlog==0.4.0
django-celery==3.2.2
django-celery-beat==1.1.0
django-celery-email==2.0.0
django-celery-results==1.0.1
django-digg-paginator==0.1.3
django-easy-pdf==0.1.1
django-extensions==1.7.5
-e git://github.com/AndrewIngram/django-extra-views.git@d9f25c316e1619c69493ea5065e808983030aca9#egg=django_extra_views
django-jsonfield==1.0.1
django-permanent==1.1.4
django-session-security==2.5.1
ephem==3.7.6.0
gunicorn==19.6.0
html5lib==0.999999999
httplib2==0.10.3
kombu==4.1.0
lxml==3.8.0
Pillow==3.4.2
pisa==3.0.33
psycopg2==2.6.2
pycparser==2.18
PyPDF2==1.26.0
Pyphen==0.9.4
pytz==2017.3
PyYAML==3.12
reportlab==2.7
six==1.10.0
tinycss2==0.5
vine==1.1.4
WeasyPrint==0.39
webencodings==0.5
Werkzeug==0.11.11
xhtml2pdf==0.2b1
